refactor: consolidate duplicate hooks into _hooks/
Phase 2a-6: 6 PR で意図的に重複させた 2 つの hook を _hooks/ に集約。

## 集約
- use-tag-creation: manager + edit → _hooks/use-tag-creation.ts
  - 重複チェックは tag.name のみ参照するため、TagLike { name: string }
    の構造的型で manager (id: string) と edit (id: number) 両方を受容
- use-tags-list: categories (axios + sort) + edit (fetch, no sort) →
  _hooks/use-tags-list.ts
  - fetch に統一、optional sortById パラメータでソート切替
  - 既存 isLoading + error logging はそのまま保持
  - categories は { sortById: true } を渡して既存挙動維持

## 削除
- app/(protected)/ems/manager/hooks/use-tag-creation.{ts,test.ts}
- app/(protected)/ems/edit/[equipmentId]/hooks/use-tag-creation.{ts,test.ts}
- app/(protected)/ems/categories/hooks/use-tags-list.{ts,test.ts}
- app/(protected)/ems/edit/[equipmentId]/hooks/use-tags-list.{ts,test.ts}

## 新規 (移動 + 統合)
- app/(protected)/ems/_hooks/use-tag-creation.{ts,test.ts}
- app/(protected)/ems/_hooks/use-tags-list.{ts,test.ts}

## 修正 (import path 更新)
- app/(protected)/ems/manager/page.tsx
- app/(protected)/ems/manager/page.test.tsx
- app/(protected)/ems/edit/[equipmentId]/page.tsx
- app/(protected)/ems/categories/page.tsx (+ sortById: true)

## 動作変更ゼロ
- categories: ソート挙動は { sortById: true } 明示で保持
- manager / edit: 関数本体は同一、structural type で型互換
- axios → fetch (categories のみ): 200 ステータスでも .reduce/.sort が
  失敗する既知パターンは変わらず ({status: 500} 本体は同じ振る舞い)
- npm test 271/271 / npm run lint 既存警告のみ / npm run build 成功
