--- app/(protected)/ems/edit/[equipmentId]/hooks/use-tag-creation.ts
@@ -1,46 +1,0 @@
-"use client";
-
-import axios from "axios";
-import { useState } from "react";
-import type { Tag as Tags } from "@/types/domain";
-
-export interface UseTagCreationParams {
-  existingTags: Tags[];
-  refetchTags: () => Promise<void>;
-}
-
-export const useTagCreation = ({ existingTags, refetchTags }: UseTagCreationParams) => {
-  const [addTagName, setAddTagName] = useState<string>("");
-  const [editTagColor, setEditTagColor] = useState<string>("");
-
-  const submit = async () => {
-    if (addTagName === "") {
-      alert("カテゴリ名は1文字以上入力してください.");
-      setAddTagName("");
-      return;
-    }
-
-    const isDuplicate = existingTags.some((tag) => tag.name === addTagName.trim());
-    if (isDuplicate) {
-      alert("このカテゴリは既に存在しています.");
-      setAddTagName("");
-      return;
-    }
-
-    await axios.post("/api/tags", {
-      name: addTagName,
-      color: editTagColor,
-    });
-    setEditTagColor("");
-    setAddTagName("");
-    await refetchTags();
-  };
-
-  return {
-    addTagName,
-    setAddTagName,
-    editTagColor,
-    setEditTagColor,
-    submit,
-  };
-};
