--- app/(protected)/ems/edit/[equipmentId]/hooks/use-tags-list.ts
@@ -1,20 +1,0 @@
-"use client";
-
-import { useEffect, useState } from "react";
-import type { Tag as Tags } from "@/types/domain";
-
-export const useTagsList = () => {
-  const [tags, setTags] = useState<Tags[]>([]);
-
-  const refetch = async () => {
-    const response = await fetch("/api/tags");
-    const data: Tags[] = await response.json();
-    setTags(data);
-  };
-
-  useEffect(() => {
-    refetch();
-  }, []);
-
-  return { tags, refetch };
-};
